servers/python/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
authlib==1.2.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pyjwt==2.8.0
cryptography==41.0.7
pymysql==1.1.0
psycopg2-binary==2.9.9
